events - Move from conmgr to common
Swap from DEBUG_FLAG_CONMGR to DEBUG_FLAG_THREAD.

Cherry-picked: c4b5905ad9
Issue: 50528
