+++ src/common/Makefile.in
@@ -165,18 +165,19 @@ libcommon_la_DEPENDENCIES = $(am__DEPENDENCIES_1)
 am_libcommon_la_OBJECTS = assoc_mgr.lo bitstring.lo callerid.lo \
 	cbuf.lo core_array.lo cpu_frequency.lo cron.lo daemonize.lo \
 	data.lo dynamic_plugin_data.lo eio.lo env.lo \
-	extra_constraints.lo fd.lo fetch_config.lo forward.lo \
-	global_defaults.lo group_cache.lo half_duplex.lo hostlist.lo \
-	http.lo http_con.lo http_router.lo http_switch.lo identity.lo \
-	id_util.lo io_hdr.lo job_features.lo job_options.lo \
-	job_record.lo job_resources.lo job_state_reason.lo list.lo \
-	log.lo msg_type.lo net.lo node_conf.lo node_features.lo \
-	oci_config.lo openapi.lo optz.lo pack.lo parse_config.lo \
-	parse_time.lo parse_value.lo part_record.lo persist_conn.lo \
-	plugin.lo plugrack.lo port_mgr.lo print_fields.lo proc_args.lo \
-	read_config.lo reverse_tree.lo run_command.lo run_in_daemon.lo \
-	sack_api.lo setproctitle.lo sluid.lo slurm_errno.lo \
-	slurm_opt.lo slurm_protocol_api.lo slurm_protocol_defs.lo \
+	extra_constraints.lo events.lo fd.lo fetch_config.lo \
+	forward.lo global_defaults.lo group_cache.lo half_duplex.lo \
+	hostlist.lo http.lo http_con.lo http_router.lo http_switch.lo \
+	identity.lo id_util.lo io_hdr.lo job_features.lo \
+	job_options.lo job_record.lo job_resources.lo \
+	job_state_reason.lo list.lo log.lo msg_type.lo net.lo \
+	node_conf.lo node_features.lo oci_config.lo openapi.lo optz.lo \
+	pack.lo parse_config.lo parse_time.lo parse_value.lo \
+	part_record.lo persist_conn.lo plugin.lo plugrack.lo \
+	port_mgr.lo print_fields.lo proc_args.lo read_config.lo \
+	reverse_tree.lo run_command.lo run_in_daemon.lo sack_api.lo \
+	setproctitle.lo sluid.lo slurm_errno.lo slurm_opt.lo \
+	slurm_protocol_api.lo slurm_protocol_defs.lo \
 	slurm_protocol_pack.lo slurm_protocol_util.lo \
 	slurm_protocol_socket.lo slurm_resolv.lo \
 	slurm_resource_info.lo slurm_rlimits_info.lo \
@@ -220,32 +221,32 @@ am__depfiles_remade = ./$(DEPDIR)/assoc_mgr.Plo \
 	./$(DEPDIR)/cpu_frequency.Plo ./$(DEPDIR)/cron.Plo \
 	./$(DEPDIR)/daemonize.Plo ./$(DEPDIR)/data.Plo \
 	./$(DEPDIR)/dynamic_plugin_data.Plo ./$(DEPDIR)/eio.Plo \
-	./$(DEPDIR)/env.Plo ./$(DEPDIR)/extra_constraints.Plo \
-	./$(DEPDIR)/fd.Plo ./$(DEPDIR)/fetch_config.Plo \
-	./$(DEPDIR)/forward.Plo ./$(DEPDIR)/global_defaults.Plo \
-	./$(DEPDIR)/group_cache.Plo ./$(DEPDIR)/half_duplex.Plo \
-	./$(DEPDIR)/hostlist.Plo ./$(DEPDIR)/http.Plo \
-	./$(DEPDIR)/http_con.Plo ./$(DEPDIR)/http_router.Plo \
-	./$(DEPDIR)/http_switch.Plo ./$(DEPDIR)/id_util.Plo \
-	./$(DEPDIR)/identity.Plo ./$(DEPDIR)/io_hdr.Plo \
-	./$(DEPDIR)/job_features.Plo ./$(DEPDIR)/job_options.Plo \
-	./$(DEPDIR)/job_record.Plo ./$(DEPDIR)/job_resources.Plo \
-	./$(DEPDIR)/job_state_reason.Plo ./$(DEPDIR)/list.Plo \
-	./$(DEPDIR)/log.Plo ./$(DEPDIR)/msg_type.Plo \
-	./$(DEPDIR)/net.Plo ./$(DEPDIR)/node_conf.Plo \
-	./$(DEPDIR)/node_features.Plo ./$(DEPDIR)/oci_config.Plo \
-	./$(DEPDIR)/openapi.Plo ./$(DEPDIR)/optz.Plo \
-	./$(DEPDIR)/pack.Plo ./$(DEPDIR)/parse_config.Plo \
-	./$(DEPDIR)/parse_time.Plo ./$(DEPDIR)/parse_value.Plo \
-	./$(DEPDIR)/part_record.Plo ./$(DEPDIR)/persist_conn.Plo \
-	./$(DEPDIR)/plugin.Plo ./$(DEPDIR)/plugrack.Plo \
-	./$(DEPDIR)/port_mgr.Plo ./$(DEPDIR)/print_fields.Plo \
-	./$(DEPDIR)/proc_args.Plo ./$(DEPDIR)/read_config.Plo \
-	./$(DEPDIR)/reverse_tree.Plo ./$(DEPDIR)/run_command.Plo \
-	./$(DEPDIR)/run_in_daemon.Plo ./$(DEPDIR)/sack_api.Plo \
-	./$(DEPDIR)/setproctitle.Plo ./$(DEPDIR)/sluid.Plo \
-	./$(DEPDIR)/slurm_errno.Plo ./$(DEPDIR)/slurm_opt.Plo \
-	./$(DEPDIR)/slurm_protocol_api.Plo \
+	./$(DEPDIR)/env.Plo ./$(DEPDIR)/events.Plo \
+	./$(DEPDIR)/extra_constraints.Plo ./$(DEPDIR)/fd.Plo \
+	./$(DEPDIR)/fetch_config.Plo ./$(DEPDIR)/forward.Plo \
+	./$(DEPDIR)/global_defaults.Plo ./$(DEPDIR)/group_cache.Plo \
+	./$(DEPDIR)/half_duplex.Plo ./$(DEPDIR)/hostlist.Plo \
+	./$(DEPDIR)/http.Plo ./$(DEPDIR)/http_con.Plo \
+	./$(DEPDIR)/http_router.Plo ./$(DEPDIR)/http_switch.Plo \
+	./$(DEPDIR)/id_util.Plo ./$(DEPDIR)/identity.Plo \
+	./$(DEPDIR)/io_hdr.Plo ./$(DEPDIR)/job_features.Plo \
+	./$(DEPDIR)/job_options.Plo ./$(DEPDIR)/job_record.Plo \
+	./$(DEPDIR)/job_resources.Plo ./$(DEPDIR)/job_state_reason.Plo \
+	./$(DEPDIR)/list.Plo ./$(DEPDIR)/log.Plo \
+	./$(DEPDIR)/msg_type.Plo ./$(DEPDIR)/net.Plo \
+	./$(DEPDIR)/node_conf.Plo ./$(DEPDIR)/node_features.Plo \
+	./$(DEPDIR)/oci_config.Plo ./$(DEPDIR)/openapi.Plo \
+	./$(DEPDIR)/optz.Plo ./$(DEPDIR)/pack.Plo \
+	./$(DEPDIR)/parse_config.Plo ./$(DEPDIR)/parse_time.Plo \
+	./$(DEPDIR)/parse_value.Plo ./$(DEPDIR)/part_record.Plo \
+	./$(DEPDIR)/persist_conn.Plo ./$(DEPDIR)/plugin.Plo \
+	./$(DEPDIR)/plugrack.Plo ./$(DEPDIR)/port_mgr.Plo \
+	./$(DEPDIR)/print_fields.Plo ./$(DEPDIR)/proc_args.Plo \
+	./$(DEPDIR)/read_config.Plo ./$(DEPDIR)/reverse_tree.Plo \
+	./$(DEPDIR)/run_command.Plo ./$(DEPDIR)/run_in_daemon.Plo \
+	./$(DEPDIR)/sack_api.Plo ./$(DEPDIR)/setproctitle.Plo \
+	./$(DEPDIR)/sluid.Plo ./$(DEPDIR)/slurm_errno.Plo \
+	./$(DEPDIR)/slurm_opt.Plo ./$(DEPDIR)/slurm_protocol_api.Plo \
 	./$(DEPDIR)/slurm_protocol_defs.Plo \
 	./$(DEPDIR)/slurm_protocol_pack.Plo \
 	./$(DEPDIR)/slurm_protocol_socket.Plo \
@@ -602,6 +603,8 @@ libcommon_la_SOURCES = \
 	env.h					\
 	extra_constraints.c			\
 	extra_constraints.h			\
+	events.c				\
+	events.h				\
 	fd.c					\
 	fd.h					\
 	fetch_config.c				\
@@ -854,6 +857,7 @@ distclean-compile:
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/dynamic_plugin_data.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/eio.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/env.Plo@am__quote@ # am--include-marker
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/events.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/extra_constraints.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/fd.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/fetch_config.Plo@am__quote@ # am--include-marker
@@ -1079,6 +1083,7 @@ distclean: distclean-am
 	-rm -f ./$(DEPDIR)/dynamic_plugin_data.Plo
 	-rm -f ./$(DEPDIR)/eio.Plo
 	-rm -f ./$(DEPDIR)/env.Plo
+	-rm -f ./$(DEPDIR)/events.Plo
 	-rm -f ./$(DEPDIR)/extra_constraints.Plo
 	-rm -f ./$(DEPDIR)/fd.Plo
 	-rm -f ./$(DEPDIR)/fetch_config.Plo
@@ -1224,6 +1229,7 @@ maintainer-clean: maintainer-clean-am
 	-rm -f ./$(DEPDIR)/dynamic_plugin_data.Plo
 	-rm -f ./$(DEPDIR)/eio.Plo
 	-rm -f ./$(DEPDIR)/env.Plo
+	-rm -f ./$(DEPDIR)/events.Plo
 	-rm -f ./$(DEPDIR)/extra_constraints.Plo
 	-rm -f ./$(DEPDIR)/fd.Plo
 	-rm -f ./$(DEPDIR)/fetch_config.Plo
