+++ src/conmgr/work.c
@@ -33,6 +33,7 @@
  *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA.
 \*****************************************************************************/
 
+#include "src/common/events.h"
 #include "src/common/list.h"
 #include "src/common/macros.h"
 #include "src/common/proc_args.h"
@@ -42,7 +43,6 @@
 
 #include "src/conmgr/conmgr.h"
 #include "src/conmgr/delayed.h"
-#include "src/conmgr/events.h"
 #include "src/conmgr/mgr.h"
 #include "src/conmgr/signals.h"
